Move upsert/delete delegation functions to `delegation-rmw-sql`
