+++ packages/delegation-process/test/integrationUtils.ts
@@ -32,6 +32,7 @@ import {
   delegationReadModelServiceBuilder,
   tenantReadModelServiceBuilder,
 } from "pagopa-interop-readmodel";
+import { upsertDelegation } from "pagopa-interop-readmodel/testUtils";
 import { delegationServiceBuilder } from "../src/services/delegationService.js";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { readModelServiceBuilderSQL } from "../src/services/readModelServiceSQL.js";
@@ -146,7 +147,7 @@ export const addOneDelegation = async (
   await writeSubmitDelegationInEventstore(delegation);
   await writeInReadmodel(delegation, delegations);
 
-  await delegationReadModelServiceSQL.upsertDelegation(delegation, 0);
+  await upsertDelegation(readModelDB, delegation, 0);
 };
 
 export const addOneTenant = async (tenant: Tenant): Promise<void> => {
