+++ packages/delegation-readmodel-writer-sql/package.json
@@ -38,6 +38,7 @@
     "pagopa-interop-commons": "workspace:*",
     "pagopa-interop-models": "workspace:*",
     "pagopa-interop-readmodel": "workspace:*",
+    "pagopa-interop-readmodel-models": "workspace:*",
     "ts-pattern": "5.2.0",
     "zod": "3.23.8"
   }
