+++ packages/delegation-readmodel-writer-sql/src/delegationConsumerServiceV2.ts
@@ -4,11 +4,11 @@ import {
   missingKafkaMessageDataError,
 } from "pagopa-interop-models";
 import { match } from "ts-pattern";
-import { ReadModelService } from "./readModelService.js";
+import { DelegationWriterService } from "./delegationWriterService.js";
 
 export async function handleMessageV2(
   message: DelegationEventEnvelopeV2,
-  readModelService: ReadModelService
+  delegationWriterService: DelegationWriterService
 ): Promise<void> {
   await match(message)
     .with(
@@ -24,7 +24,7 @@ export async function handleMessageV2(
         if (!message.data.delegation) {
           throw missingKafkaMessageDataError("delegation", message.type);
         }
-        await readModelService.upsertDelegation(
+        await delegationWriterService.upsertDelegation(
           fromDelegationV2(message.data.delegation),
           message.version
         );
