+++ packages/readmodel-checker/test/delegationReadModelCheck.test.ts
@@ -7,10 +7,12 @@ import {
   WithMetadata,
   delegationKind,
 } from "pagopa-interop-models";
+import { upsertDelegation } from "pagopa-interop-readmodel/testUtils";
 import { compare } from "../src/utils.js";
 import {
   addOneDelegation,
   delegationReadModelServiceSQL,
+  readModelDB,
   readModelService,
   readModelServiceSQL,
 } from "./utils.js";
@@ -49,7 +51,8 @@ describe("Check delegation readmodels", () => {
 
     await addOneDelegation(delegation);
 
-    await delegationReadModelServiceSQL.upsertDelegation(
+    await upsertDelegation(
+      readModelDB,
       delegation.data,
       delegation.metadata.version
     );
@@ -83,7 +86,8 @@ describe("Check delegation readmodels", () => {
     await addOneDelegation(delegation1);
     await addOneDelegation(delegation2);
 
-    await delegationReadModelServiceSQL.upsertDelegation(
+    await upsertDelegation(
+      readModelDB,
       delegation2.data,
       delegation2.metadata.version
     );
@@ -116,11 +120,13 @@ describe("Check delegation readmodels", () => {
 
     await addOneDelegation(delegation1);
 
-    await delegationReadModelServiceSQL.upsertDelegation(
+    await upsertDelegation(
+      readModelDB,
       delegation1.data,
       delegation1.metadata.version
     );
-    await delegationReadModelServiceSQL.upsertDelegation(
+    await upsertDelegation(
+      readModelDB,
       delegation2.data,
       delegation2.metadata.version
     );
@@ -156,7 +162,8 @@ describe("Check delegation readmodels", () => {
 
     await addOneDelegation(delegation1);
 
-    await delegationReadModelServiceSQL.upsertDelegation(
+    await upsertDelegation(
+      readModelDB,
       delegation1InPostgresDb.data,
       delegation1InPostgresDb.metadata.version
     );
@@ -191,7 +198,8 @@ describe("Check delegation readmodels", () => {
 
     await addOneDelegation(delegation1);
 
-    await delegationReadModelServiceSQL.upsertDelegation(
+    await upsertDelegation(
+      readModelDB,
       delegation1InPostgresDb.data,
       delegation1InPostgresDb.metadata.version
     );
