+++ packages/readmodel/src/testUtils.ts
@@ -1,16 +1,23 @@
 import {
+  Delegation,
   TenantNotificationConfig,
   UserNotificationConfig,
 } from "pagopa-interop-models";
 import {
+  delegationContractDocumentInReadmodelDelegation,
+  delegationInReadmodelDelegation,
+  delegationStampInReadmodelDelegation,
   DrizzleReturnType,
   tenantNotificationConfigInReadmodelNotificationConfig,
   userNotificationConfigInReadmodelNotificationConfig,
 } from "pagopa-interop-readmodel-models";
+import { eq } from "drizzle-orm";
 import {
   splitTenantNotificationConfigIntoObjectsSQL,
   splitUserNotificationConfigIntoObjectsSQL,
 } from "./notification-config/splitters.js";
+import { splitDelegationIntoObjectsSQL } from "./delegation/splitters.js";
+import { checkMetadataVersion } from "./utils.js";
 
 export const insertTenantNotificationConfig = async (
   readModelDB: DrizzleReturnType,
@@ -40,4 +47,42 @@ export const insertUserNotificationConfig = async (
         metadataVersion
       )
     );
+};
+
+export const upsertDelegation = async (
+  readModelDB: DrizzleReturnType,
+  delegation: Delegation,
+  metadataVersion: number
+): Promise<void> => {
+  await readModelDB.transaction(async (tx) => {
+    const shouldUpsert = await checkMetadataVersion(
+      tx,
+      delegationInReadmodelDelegation,
+      metadataVersion,
+      delegation.id
+    );
+
+    if (!shouldUpsert) {
+      return;
+    }
+
+    await tx
+      .delete(delegationInReadmodelDelegation)
+      .where(eq(delegationInReadmodelDelegation.id, delegation.id));
+
+    const { delegationSQL, stampsSQL, contractDocumentsSQL } =
+      splitDelegationIntoObjectsSQL(delegation, metadataVersion);
+
+    await tx.insert(delegationInReadmodelDelegation).values(delegationSQL);
+
+    for (const stampSQL of stampsSQL) {
+      await tx.insert(delegationStampInReadmodelDelegation).values(stampSQL);
+    }
+
+    for (const docSQL of contractDocumentsSQL) {
+      await tx
+        .insert(delegationContractDocumentInReadmodelDelegation)
+        .values(docSQL);
+    }
+  });
 };
